Problem: RtpsRelay must be manually configured
The RtpsRelay instance in GCP changes IP address every time that it is
restarted.  Consequently, the person running the demo must know where
to inject the IP address for the demo to work correctly.

Solution: Add the RTPS_RELAY_IP environment variable to the
configuration file and inject it into the OpenDDS configuration for
both the Raspberry Pi and the Android app.  The configuration file for
the Android app is a static asset that is compiled.  Using DNS would
eliminate the need to inject the IP.