Problem: Android app is configured with list of locks to expect
The person running the demo should have the freedom to use whatever
lock ids they choose and expect them to appear in the Android app.

Solution: Read in the expected lock ids from the configuration file
and inject them into the configuration for the Android app before it
is compiled.