Problem: RtpsRelay fills disk with logs and uses custom Docker image
Solution: Use the standard OpenDDS image in Dockerhub and disable
logging in the Docker daemon.  The logs are still accessible through
journalctl which should also rotate them correctly.

Note that the Docker image for the relay is tagged with master.  An
image tagged with a version should be used once one that has the
RtpsRelay build is available.